backend/requirements.txt
asgiref==3.8.1
Django~=5.2.4
sqlparse==0.5.3
tzdata>=2024.1
pytest==8.3.3
pytest-django==4.8.0
PyJWT==2.10.1
djangorestframework~=3.16.0
django-cors-headers>=4.3,<5
djoser>=2.2,<3
djangorestframework-simplejwt>=5.3,<6

psycopg[binary]>=3.1